# book-network-ui/src/app/services/models/reset-password-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ResetPasswordRequest {
  confirmPassword?: string;
  newPassword?: string;
  token?: string;
}